Add DiffOperationAccordionTable component and integrate it into TimelineRecordDiffSnapshotEvent; enhance utils for formatting values and paths; update styles for record-diff-table
