+++ oarepo_requests/ui/theme/assets/semantic-ui/js/oarepo_requests_ui/common/components/TimelineRecordDiffSnapshotEvent.jsx
@@ -1,30 +1,33 @@
-import React, { useState } from "react";
+import React from "react";
 import PropTypes from "prop-types";
-import { Icon, Table, Message, Accordion } from "semantic-ui-react";
+import { Icon, Message } from "semantic-ui-react";
 import { i18next } from "@translations/oarepo_requests_ui/i18next";
 import { ErrorBoundary } from "react-error-boundary";
 import _isArray from "lodash/isArray";
-import _isObjectLike from "lodash/isObject";
-import _every from "lodash/every";
-import _isString from "lodash/isString";
 import {
   getRequestStatusIcon,
   getFeedMessage,
   GenericActionEvent,
+  DiffOperationAccordionTable,
 } from "@js/oarepo_requests_common";
 
-export const TimelineRecordDiffSnapshotEvent = ({ event }) => {
-  const [visibleAccordions, setVisibleAccordions] = useState({});
+const DiffFallbackMessage = ({ error }) => (
+  <Message negative>
+    <Message.Header>
+      {i18next.t("Unable to parse diff data")}
+    </Message.Header>
+    <p>{i18next.t("There was an error processing the diff data.")}</p>
+    {error?.message && <pre>{error.message}</pre>}
+  </Message>
+);
 
+DiffFallbackMessage.propTypes = {
+  error: PropTypes.object.isRequired,
+};
+
+export const TimelineRecordDiffSnapshotEvent = ({ event }) => {
   const eventIcon = getRequestStatusIcon("edited");
   const feedMessage = getFeedMessage("edited");
-
-  const toggleAccordion = (operationType) => {
-    setVisibleAccordions(prev => ({
-      ...prev,
-      [operationType]: !prev[operationType]
-    }));
-  };
 
   // Parse and process diff data
   const renderDiffTables = () => {
@@ -56,130 +59,25 @@ export const TimelineRecordDiffSnapshotEvent = ({ event }) => {
       return groups;
     }, {});
 
-    // Format values for display
-    const formatValue = (value) => {
-      if (value === null || value === undefined) return "—";
-      if (_isArray(value) && _every(value, _isString)) return value.join(", ");
-      if (_isObjectLike(value))
-        return JSON.stringify(value, null, 2);
-      return String(value);
-    };
-
-    // Convert path to human readable format
-    const formatPath = (path) => {
-      return path
-        .replace(/^\//, "")
-        .replace(/\/(\d+)\//g, (match, arrayIndex) => {
-          return ` › ${parseInt(arrayIndex) + 1} › `;
-        })
-        .replace(/\/(\d+)$/, (match, arrayIndex) => {
-          return ` › ${parseInt(arrayIndex) + 1}`;
-        })
-        .replace(/\//g, " › ");
-    };
-
-    const renderOperationTable = (operations, operationType) => {
-      const operationConfig = {
-        add: {
-          title: i18next.t("Added"),
-          icon: "add",
-          color: "green"
-        },
-        remove: {
-          title: i18next.t("Removed"),
-          icon: "remove",
-          color: "red"
-        },
-        replace: {
-          title: i18next.t("Changed"),
-          icon: "pencil",
-          color: "orange"
-        }
-      };
-
-      const config = operationConfig[operationType];
-      if (!config || operations.length === 0) return null;
-
-      const isVisible = visibleAccordions[operationType] || false;
-
-      return (
-        <Accordion key={operationType} fluid className={`operation-accordion ${operationType} rel-mb-1`}>
-          <Accordion.Title
-            active={isVisible}
-            onClick={() => toggleAccordion(operationType)}
-            className={`operation-title ${operationType} flex align-items-center`}
-          >
-            <Icon name="dropdown" />
-            <span>
-              <Icon name={config.icon} className="operation-icon" />
-              {config.title}
-              <span className="operation-count">({operations.length})</span>
-            </span>
-          </Accordion.Title>
-          <Accordion.Content active={isVisible}>
-            <Table basic celled stackable className={`record-diff-table requests ${operationType} borderless`}>
-              <Table.Header>
-                <Table.Row>
-                  <Table.HeaderCell>{i18next.t("Field")}</Table.HeaderCell>
-                  {operationType === "replace" ? (
-                    <>
-                      <Table.HeaderCell>{i18next.t("Old Value")}</Table.HeaderCell>
-                      <Table.HeaderCell>{i18next.t("New Value")}</Table.HeaderCell>
-                    </>
-                  ) : (
-                    <Table.HeaderCell>{i18next.t("Value")}</Table.HeaderCell>
-                  )}
-                </Table.Row>
-              </Table.Header>
-              <Table.Body>
-                {operations.map((op, index) => (
-                  <Table.Row key={`${operationType}-${index}`}>
-                    <Table.Cell>
-                      <code>{formatPath(op.path)}</code>
-                    </Table.Cell>
-                    {operationType === "replace" ? (
-                      <>
-                        <Table.Cell>
-                          <pre>{formatValue(op.old_value)}</pre>
-                        </Table.Cell>
-                        <Table.Cell>
-                          <pre>{formatValue(op.value)}</pre>
-                        </Table.Cell>
-                      </>
-                    ) : (
-                      <Table.Cell>
-                        <pre>{formatValue(operationType === "remove" ? op.old_value : op.value)}</pre>
-                      </Table.Cell>
-                    )}
-                  </Table.Row>
-                ))}
-              </Table.Body>
-            </Table>
-          </Accordion.Content>
-        </Accordion>
-      );
-    };
-
     return (
       <ErrorBoundary FallbackComponent={DiffFallbackMessage}>
         <div className="diff-tables-container ml-5">
-          {renderOperationTable(operationsByType.add || [], "add")}
-          {renderOperationTable(operationsByType.remove || [], "remove")}
-          {renderOperationTable(operationsByType.replace || [], "replace")}
+          <DiffOperationAccordionTable
+            operations={operationsByType.add || []}
+            operationType="add"
+          />
+          <DiffOperationAccordionTable
+            operations={operationsByType.remove || []}
+            operationType="remove"
+          />
+          <DiffOperationAccordionTable
+            operations={operationsByType.replace || []}
+            operationType="replace"
+          />
         </div>
       </ErrorBoundary>
     );
   };
-
-  const DiffFallbackMessage = ({ error }) => (
-    <Message negative>
-      <Message.Header>
-        {i18next.t("Unable to parse diff data")}
-      </Message.Header>
-      <p>{i18next.t("There was an error processing the diff data.")}</p>
-      {error?.message && <pre>{error.message}</pre>}
-    </Message>
-  );
 
   return (
     <GenericActionEvent
