+++ oarepo_requests/ui/theme/assets/semantic-ui/js/oarepo_requests_ui/common/components/index.js
@@ -25,3 +25,4 @@ export * from "./ConfirmationModal";
 export * from "./TimelineEscalationEvent";
 export * from "./GenericActionEvent";
 export * from "./TimelineRecordDiffSnapshotEvent";
+export * from "./DiffOperationAccordionTable";
