+++ oarepo_requests/ui/theme/assets/semantic-ui/less/oarepo_requests_ui/themes/default/views/feed.overrides
@@ -101,7 +101,7 @@
                         
             .record-diff-table {
               margin-top: 0;
-              margin-left: 0.5rem;
+              padding-left: 0.5rem;
               width: 100%;
               word-break: normal;
